refactor: separate single types module into several ones
